Update Jenkinsfile
test

+++ Jenkinsfile
@@ -42,15 +42,17 @@ pipeline {
        steps {
         withSonarQubeEnv('CB2Sonarrehl8') {
           script {
-            def readContent = readFile "sonar-project.properties"
-            writeFile file: "sonar-project.properties", text: "$readContent \nsonar.branch.name=$BRANCH_NAME\n"
-            docker.image("${sonarImage}").withRun('--security-opt label=disable -v "$PWD:/usr/src"') { c ->
+            if (env.GIT_BRANCH == 'dev' || env.GIT_BRANCH == 'master' ) {
+                  def readContent = readFile "sonar-project.properties"
+                  writeFile file: "sonar-project.properties", text: "$readContent \nsonar.branch.name=$BRANCH_NAME\n"
+                  docker.image("${sonarImage}").withRun('--security-opt label=disable -v "$PWD:/usr/src"') { c ->
             
-            // NEED THIS LINE TO WORK!
-            sh 'while [ ! -f ./.scannerwork/report-task.txt ]; do sleep 5; done'
-            sh 'sleep 10'
-            sh 'if [ -d ./.scannerwork ]; then chmod -R 777 ./.scannerwork; fi'
-            sh 'if [ -d ./.scannerwork ]; then rm -Rf ./.scannerwork; fi'
+                 // NEED THIS LINE TO WORK!
+                 sh 'while [ ! -f ./.scannerwork/report-task.txt ]; do sleep 5; done'
+                 sh 'sleep 10'
+                 sh 'if [ -d ./.scannerwork ]; then chmod -R 777 ./.scannerwork; fi'
+                 sh 'if [ -d ./.scannerwork ]; then rm -Rf ./.scannerwork; fi'
+              }     
             }
           }
         }
